ai_panel/requirements.txt
flask>=3.0
openai>=1.50
google-generativeai>=0.8.3
anthropic>=0.40
python-dotenv>=1.0
